rename cookie utility
