+++ packages/identity-x/browser/utils/cookies-enabled.js
